Changed weight to double

+++ src/Svea.WebPay.SDK/CheckoutApi/GetOrderShippingInformation.cs
@@ -13,12 +13,12 @@
         /// <param name="tags"></param>
         /// <param name="fallbackOptions"></param>
         /// <param name="shippingProvider"></param>
-        public GetOrderShippingInformation(bool enforceFallback, bool enableShipping, int weight, Dictionary<string, string> tags, ShippingOption fallbackOptions, ShippingProvider shippingProvider)
+        public GetOrderShippingInformation(bool enforceFallback, bool enableShipping, double weight, Dictionary<string, string> tags, ShippingOption fallbackOptions, ShippingProvider shippingProvider)
             : base(enforceFallback, enableShipping, weight, tags, fallbackOptions)
         {
             ShippingProvider = shippingProvider;
         }
-        
+
         public ShippingProvider ShippingProvider { get; }
     }
 }
